--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         version:
-          - 1.80.0
+          - 1.82.0
           - stable
           - beta
           - nightly