--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.2]
+### Changed
+- More performance improvements this time changing search order to prefer deeper expression nodes first
+
 ## [0.8.1] 2025-05-06
 ### Changed
 - Increased usage of memoization of hashes -> profiles to speedup report generation more